Mess

+++ src/uu/dd/src/datastructures.rs
@@ -131,8 +131,8 @@ pub enum StatusLevel {
 /// then becomes Bytes(N)
 #[derive(Debug, PartialEq)]
 pub enum CountType {
-    Reads(usize),
-    Bytes(usize),
+    Reads(u64),
+    Bytes(u64),
 }
 
 #[derive(Debug)]
